refactor: contexts and builders
introduces the NotificationContext and ContextBuilder classes.
the only specialization currently is for the PR comment.

the context builders generate the context details necessary to create and
send the message later on. Each step is called in order, and if one step fails
we bail with an exception. This means the notification can't be sent.